==> R/dev/06_FirstModelling.Rmd <==
---
title: "Next Word Modelling (week 2)"
author: "BB"
date: '2022-04-28'
output: html_document
---

```{r setup, include=FALSE}
# Libraries
library(stringr)
library(tidyverse)
library(caret)
library(recipes)


# Parameters
knitr::opts_chunk$set(echo = TRUE)

# Data 
word <- read_rds("../data/en_US.twitter_word.rds.gz")
gram2 <- read_rds("../data/en_US.twitter_gram2.rds.gz")
gram3 <- read_rds("../data/en_US.twitter_gram3.rds.gz")
gram4 <- read_rds("../data/en_US.twitter_gram4.rds.gz")

# Functions

source("./helper_functions.R")


```

## 1. Dataset Building

Datasets needed for a basic n-gram model for predicting the next word based on the previous 1, 2, or 3 words. Only the top 100 words survive in any n-gram.

```{r datasets}

swiftkey_1 <- as.data.frame(word) %>% 
    group_by(word) %>% 
    summarize(freq = n()) %>% 
    arrange(-freq) %>% 
    filter(word != "") %>%
    head(1000)
    

swiftkey_2 <- as.data.frame(gram2) %>% 
    separate(gram2, c("w1", "w2"), " ") %>% 
    filter(w1 %in% swiftkey_1$word,
           w2 %in% swiftkey_1$word) %>% 
    group_by(w1,w2) %>% 
    summarize(freq = n()) %>% 
    arrange(-freq)

swiftkey_3 <- as.data.frame(gram3) %>% 
    separate(gram3, c("w1", "w2", "w3"), " ") %>% 
    filter(w1 %in% swiftkey_1$word,
           w2 %in% swiftkey_1$word,
           w3 %in% swiftkey_1$word) %>% 
    group_by(w1,w2,w3) %>% 
    summarize(freq = n()) %>% 
    arrange(-freq)

swiftkey_4 <- as.data.frame(gram4) %>% 
    separate(gram4, c("w1", "w2", "w3", "w4"), " ") %>% 
    filter(w1 %in% swiftkey_1$word,
           w2 %in% swiftkey_1$word,
           w3 %in% swiftkey_1$word,
           w4 %in% swiftkey_1$word) %>% 
    group_by(w1,w2, w3, w4) %>% 
    summarize(freq = n()) %>% 
    arrange(-freq)

swiftkey <- list("1" = swiftkey_1,
                 "2" = swiftkey_2,
                 "3" = swiftkey_3,
                 "4" = swiftkey_4)


```


## 2. Prediction

Predict next word from one given word, based on bigrams:

```{r bigram prediction}

predict2 <- function(corpus, word1) {

    set <- corpus[corpus$w1 == word1,]
    
    n_cases <- sum(set$freq, na.rm = TRUE)
    
    probs <- set %>% 
        mutate(p = round(freq / n_cases,2)) %>% 
        select(-freq) %>% 
        arrange(-p)
    
    probs$w2[1]
}

predict2(swiftkey_2, "you")
predict2(swiftkey_2, "to")
predict2(swiftkey_2, "me")
predict2(swiftkey_2, "oger")


```

Predict next words from a given two-word phrase based on trigrams:

```{r trigram prediction}

predict3 <- function(corpus, word1, word2) {
    
    set <- corpus %>% 
        filter(w1==word1, w2==word2)
    
    n_cases <- sum(set$freq, na.rm = TRUE)
    
    probs <- set %>% 
        mutate(p = round(freq / n_cases,2)) %>% 
        select(-freq)
    
    probs$w3[1]
}

predict3(swiftkey_3, "don't", "you")
predict3(swiftkey_3, "love", "to")
predict3(swiftkey_3, "make", "me")
predict3(swiftkey_3, "big", "oger")


```

Predict the next words form a given three-word phrase based on tetragrams:

```{r tetragram prediction}

predict4 <- function(corpus, word1, word2, word3) {
    
    set <- corpus %>% 
        filter(w1==word1, w2==word2, w3 == word3)
    
    n_cases <- sum(set$freq, na.rm = TRUE)
    
    probs <- set %>% 
        mutate(p = round(freq / n_cases,2)) %>% 
        select(-freq)
    
    probs$w4[1]

}

predict4(swiftkey_4,"why", "don't", "you")
predict4(swiftkey_4,"i", "love", "to")
predict4(swiftkey_4,"you", "make", "me")
predict4(swiftkey_4,"the", "big", "oger")


```

Predict from a given phrase by combination of above and handle unseen ngrams:

```{r predict phrase}

predict_phrase <- function(corpus, phrase) {
    
    # corpus <- swiftkey
    # phrase <- "big"
    # 
    clean_phrase <- clean_text(phrase)
    final_words <- last_n_words(clean_phrase, 3)
    
    pred <- NA
    
    if(length(final_words) == 3) 
        pred <- predict4(corpus[["4"]], 
                     final_words[1], 
                     final_words[2], 
                     final_words[3])
    if(length(final_words) == 2) 
        pred <- predict3(corpus[["3"]], 
                     final_words[1], 
                     final_words[2])
    if(length(final_words) == 1)
        pred <- predict2(corpus[["2"]], 
                     final_words[1])
    
    pred
    
    if(is.na(pred)) corpus[["1"]]$word[1] else pred
}

predict_phrase(swiftkey, "me")
predict_phrase(swiftkey, "make me")
predict_phrase(swiftkey, "you make me")
predict_phrase(swiftkey, "the big oger")
predict_phrase(swiftkey, "I thank you")

# Week 3 Quiz:
predict_phrase(swiftkey, "and a case of")
predict_phrase(swiftkey, "It would mean the")
predict_phrase(swiftkey, "and make me the")
predict_phrase(swiftkey, "still struggling but the")
predict_phrase(swiftkey, "romantic date at the")
predict_phrase(swiftkey, "them off and be on my")
predict_phrase(swiftkey, "and be on my")
predict_phrase(swiftkey, "seen it in quite some")
predict_phrase(swiftkey, "his eyes with his little")
predict_phrase(swiftkey, "the faith during the")
predict_phrase(swiftkey, "then you must be")

```

# The KI Story

```{r KI story}

phrase <- "i love to"

for(i in 1:10) {
    new <- predict_phrase(swiftkey, phrase)
    phrase <- paste(phrase,new)
}

predict_phrase(swiftkey, "i love to")

```
